requirements.txt
opencv-python==4.5.1.48
paho-mqtt==1.5.1